chore: created new `GradeCompositionOverlayComponent`
- moved `<.grades_report_grid>` and `<.students_grades_grid>` from `Reporting` to `GradesReports` "context"

- implemented `GradeCompositionOverlayComponent` in `GradesReportsLive`
