+++ lib/lanttern_web/live/pages/grading/grades_reports_live.ex
@@ -9,10 +9,11 @@ defmodule LantternWeb.GradesReportsLive do
 
   # live components
   alias LantternWeb.GradesReports.GradesReportFormComponent
+  alias LantternWeb.Grading.GradeCompositionOverlayComponent
 
   # shared
+  import LantternWeb.GradesReportsComponents
   import LantternWeb.GradingComponents
-  import LantternWeb.ReportingComponents
 
   # lifecycle
 
@@ -35,6 +36,7 @@ defmodule LantternWeb.GradesReportsLive do
       |> assign(:has_grades_reports, length(grades_reports) > 0)
       |> assign_show_grades_report_form(params)
       |> assign_show_grades_report_grid_editor(params)
+      |> assign_is_editing_grade_composition(params)
 
     {:noreply, socket}
   end
@@ -90,6 +92,24 @@ defmodule LantternWeb.GradesReportsLive do
   defp assign_show_grades_report_grid_editor(socket, _),
     do: assign(socket, :show_grades_report_grid_editor, false)
 
+  defp assign_is_editing_grade_composition(socket, %{
+         "gr_id" => grades_report_id,
+         "grc_id" => grades_report_cycle_id,
+         "grs_id" => grades_report_subject_id
+       }) do
+    socket
+    |> assign(:grades_report_id, grades_report_id)
+    |> assign(:grades_report_cycle_id, grades_report_cycle_id)
+    |> assign(:grades_report_subject_id, grades_report_subject_id)
+  end
+
+  defp assign_is_editing_grade_composition(socket, _) do
+    socket
+    |> assign(:grades_report_id, nil)
+    |> assign(:grades_report_cycle_id, nil)
+    |> assign(:grades_report_subject_id, nil)
+  end
+
   # event handlers
 
   @impl true
@@ -110,5 +130,19 @@ defmodule LantternWeb.GradesReportsLive do
 
         {:noreply, socket}
     end
+  end
+
+  def handle_event("edit_composition", params, socket) do
+    url_params = %{
+      gr_id: params["gradesreportid"],
+      grc_id: params["gradesreportcycleid"],
+      grs_id: params["gradesreportsubjectid"]
+    }
+
+    socket =
+      socket
+      |> push_patch(to: ~p"/grading?#{url_params}")
+
+    {:noreply, socket}
   end
 end
